requirements.txt
streamlit
requests
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu
